추상팩터리 패턴
