+++ java.study/src/main/java/designpattern/creation/factory/abstractfactory/ex2/Ingredient/SlicedPepperoni.java
@@ -1,6 +1,6 @@
 package designpattern.creation.factory.abstractfactory.ex2.Ingredient;
 
-import designpattern.creation.factory.abstractfactory.ex2.Pepperoni;
+import designpattern.creation.factory.abstractfactory.ex2.product.Pepperoni;
 
 /**
  *
