+++ java.study/src/main/java/designpattern/creation/factory/abstractfactory/ex2/client/NYPizzaStore.java
@@ -1,23 +1,26 @@
-package designpattern.creation.factory.abstractfactory.ex2.product;
+package designpattern.creation.factory.abstractfactory.ex2.client;
 
-import designpattern.creation.factory.abstractfactory.ex2.Pizza;
 import designpattern.creation.factory.abstractfactory.ex2.concrate.CheesePizza;
+import designpattern.creation.factory.abstractfactory.ex2.product.Pizza;
 import designpattern.creation.factory.abstractfactory.ex2.concrate.PepperoniPizza;
 import designpattern.creation.factory.abstractfactory.ex2.concrate.VeggiePizza;
-import designpattern.creation.factory.abstractfactory.ex2.factory.ChicagoIngredientFactory;
+import designpattern.creation.factory.abstractfactory.ex2.factory.NYPizzaIngredientFactory;
 import designpattern.creation.factory.abstractfactory.ex2.factory.PizzaIngredientFactory;
 
 /**
+ * <pre>
+ * Description :
  *
- *  *  @author skan
+ *
+ * </pre>
+ *
+ * @author skan
+ * @version Copyright (C) 2022 by CJENM|MezzoMedia. All right reserved.
  * @since 2022/12/14
  */
-public class ChicagoPizzaStore extends PizzaStore{
+public class NYPizzaStore extends PizzaStore {
 
-
-    // 피자 생성시 어떤 재료공장의 팩토리를 사용할지 선택한다.
-    final PizzaIngredientFactory pizzaIngredientFactory = new ChicagoIngredientFactory();
-
+    PizzaIngredientFactory pizzaIngredientFactory = new NYPizzaIngredientFactory();
 
     @Override
     Pizza createPizza(String name) {
@@ -28,11 +31,10 @@ public class ChicagoPizzaStore extends PizzaStore{
         } else if (name.equals("veggie")) {
             pizza = new VeggiePizza(pizzaIngredientFactory);
         } else if (name.equals("pepperoni")) {
-            pizza =  new PepperoniPizza(pizzaIngredientFactory);
+            pizza = new PepperoniPizza(pizzaIngredientFactory);
         } else {
             throw new IllegalArgumentException();
         }
-
         return pizza;
     }
 }
