+++ java.study/src/main/java/designpattern/creation/factory/abstractfactory/ex2/factory/ChicagoIngredientFactory.java
@@ -1,7 +1,7 @@
 package designpattern.creation.factory.abstractfactory.ex2.factory;
 
-import designpattern.creation.factory.abstractfactory.ex2.*;
 import designpattern.creation.factory.abstractfactory.ex2.Ingredient.*;
+import designpattern.creation.factory.abstractfactory.ex2.product.*;
 
 /**
  *
